fix: implement printing; cleanup options toggle
