+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2024/01/20 13:44:44 by gwolf            ###   ########.fr        #
+#    Updated: 2024/01/20 14:07:27 by gwolf            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -190,7 +190,6 @@ SRC_COMMON := 	camera_movement.c \
 				render_keyhook_colour.c \
 				render_keyhook_hittable.c \
 				render_keyhook_light.c \
-				render_keyhook_press.c \
 				render_keyhook_scene.c \
 				render_keyhook_utils.c \
 				render_loop_mlx.c \
@@ -213,6 +212,7 @@ SRC_BASE := 	colour.c \
 				render_draw.c \
 				render_init_base.c \
 				render_keyhook_options.c \
+				render_keyhook_press.c \
 				scene_light.c \
 				utils_colour.c \
 				utils_cylinder.c \
@@ -236,7 +236,9 @@ SRC_BONUS :=	check_bonus.c \
 				render_draw_bonus.c \
 				render_init_bonus.c \
 				render_keyhook_options_bonus.c \
+				render_keyhook_press_bonus.c \
 				render_output_ppm_bonus.c \
+				render_output_threaded.c \
 				scene_light_bonus.c \
 				scene_reflection_bonus.c \
 				texture_bonus.c \
