+++ src/render_keyhook_press.c
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/19 22:40:44 by gwolf             #+#    #+#             */
-/*   Updated: 2024/01/19 17:28:49 by gwolf            ###   ########.fr       */
+/*   Updated: 2024/01/20 14:00:59 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 static bool	ft_is_option_key(int key)
 {
 	return (key == XK_Escape || key == XK_Shift_L || key == XK_Control_L
-		|| key == XK_i || key == XK_o || key == XK_u);
+		|| key == XK_i || key == XK_u);
 }
 
 static bool	ft_is_manip_key(int key)
@@ -31,8 +31,6 @@ int	ft_keyhook_press(int key, t_render *render)
 {
 	float	inc;
 
-	if (key == XK_Escape)
-		return (mlx_loop_end(render->mlx_ptrs.mlx_ptr));
 	if (ft_is_option_key(key))
 		ft_change_options(key, render);
 	else if (ft_is_manip_key(key))
