+++ src/render_keyhook_press_bonus.c
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/19 22:40:44 by gwolf             #+#    #+#             */
-/*   Updated: 2024/01/19 17:30:18 by gwolf            ###   ########.fr       */
+/*   Updated: 2024/01/20 14:07:54 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,8 @@
 static bool	ft_is_option_key(int key)
 {
 	return (key == XK_Escape || key == XK_Shift_L || key == XK_Control_L
-		|| key == XK_i || key == XK_o || key == XK_u);
+		|| key == XK_i || key == XK_o || key == XK_u || key == XK_p
+		|| key == XK_j);
 }
 
 static bool	ft_is_manip_key(int key)
@@ -29,26 +30,10 @@ static bool	ft_is_manip_key(int key)
 
 int	ft_keyhook_press(int key, t_render *render)
 {
-	if (ft_is_printing(render))
-		return (0);
-	else if (key == XK_Escape)
-		return (mlx_loop_end(render->mlx_ptrs.mlx_ptr));
-	if (key == XK_p)
-	{
-		ft_toggle_is_printing(render);
-		ft_spin_detached_thread(render, ft_output_threaded);
-		return (0);
-	}
-	else if (key == XK_j)
-		render->is_threaded = !render->is_threaded;
-	else if (key == XK_i)
-		ft_toggle_menu(render);
-	else if (key == XK_Control_L)
-		ft_change_mode(render);
-	else if (key == XK_Shift_L)
-		ft_change_menu_page_ctrl(render);
 	float	inc;
 
+	if (ft_is_printing(render))
+		return (0);
 	if (ft_is_option_key(key))
 		ft_change_options(key, render);
 	else if (ft_is_manip_key(key))
@@ -59,7 +44,7 @@ int	ft_keyhook_press(int key, t_render *render)
 		else if (ft_option_isset(render->options, O_MODE_LIGHT))
 			ft_manip_light(key, &render->scene, &render->active_light, inc);
 		else if (ft_option_isset(render->options, O_MODE_CAM))
-			ft_manip_cam(key, &render->cam, inc);
+			ft_manip_cam(key, &render->cam, inc, &render->options);
 		ft_option_set(&render->options, O_SCENE_CHANGED);
 	}
 	return (0);
