+++ src/threads_alt_versions_bonus.c
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/14 10:14:59 by gwolf             #+#    #+#             */
-/*   Updated: 2024/01/19 18:33:14 by gwolf            ###   ########.fr       */
+/*   Updated: 2024/01/20 13:52:50 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
